项目已暂时完成

+++ Graduate/src/main/java/com/txy/graduate/service/Impl/ProcessServiceImpl.java
@@ -32,6 +32,8 @@ public class ProcessServiceImpl extends ServiceImpl<ProcessMapper, Process> impl
     @Override
     public Process queryById(Long id) {
         Process process = mapper.selectById(id);
+        if(process==null)
+            return null;
         //对查询到的结果进行处理
         return fixData(process).get(0);
     }
@@ -40,7 +42,7 @@ public class ProcessServiceImpl extends ServiceImpl<ProcessMapper, Process> impl
     public boolean saveProcess(Process process) {
         //为数据添加图片存放的根路径
         String baseImgPath = fileUtil.getRootPath(Process.class);
-        if(baseImgPath==null||baseImgPath.equals("")){
+        if (baseImgPath == null || baseImgPath.equals("")) {
             return false;
         }
         //设置图片存放的根目录
@@ -57,33 +59,38 @@ public class ProcessServiceImpl extends ServiceImpl<ProcessMapper, Process> impl
         //查询新的process中的imgPaths
         List<String> imgPaths = process.getImgPaths();
         //查询旧的process中的imgPaths
-        List<String> imgPathsOld = queryById(process.getId()).getImgPaths();
+        Process query = queryById(process.getId());
+        if(query==null)
+            return false;
+        List<String> imgPathsOld = query.getImgPaths();
+
         //删除的旧的图片
         imgPathsOld.forEach(i->{
-            if(!imgPaths.contains(i)){
+            if (!imgPaths.contains(i))
                 fileUtil.removeFiles(i);
-            }
         });
 
-        return mapper.updateById(joinData(process).get(0))>0;
+        //执行更新数据库数据
+        return mapper.updateById(process) > 0;
     }
 
     @Override
     public boolean removeProcessById(Long id) {
         //查询对应的process，获取到对应的图片路径信息
         Process process = queryById(id);
+        if(process==null)
+            return false;
         //删除这些图片
         fileUtil.removeFiles(process.getImgPaths().toArray(String[]::new));
-
         //删除数据库中的数据
-        return mapper.deleteById(id)>0;
+        return mapper.deleteById(id) > 0;
     }
 
 
     //对数据库中的数据修复成完整的数据
     public List<Process> fixData(Process... processes) {
-        if(processes!=null && processes.length>0){
-            return Arrays.stream(processes).peek(process->{
+        if (processes[0] != null) {
+            return Arrays.stream(processes).peek(process -> {
                 //获取存放的根路径
                 String baseImgPath = process.getBaseImgPath();
                 //补全图片的完整路径
@@ -104,7 +111,7 @@ public class ProcessServiceImpl extends ServiceImpl<ProcessMapper, Process> impl
 
     //对前端发来的数据进行拼接，img[] --> img
     public List<Process> joinData(Process... processes) {
-        return Arrays.stream(processes).map(p->{
+        return Arrays.stream(processes).map(p -> {
             //获取存放的根路径
             String baseImgPath = p.getBaseImgPath();
             List<String> imgPaths = p.getImgPaths();
